make ImpactTdContents/ThContents into components
