+++ src/packages/@ncigdc/components/Impacts/index.js
@@ -1,0 +1,4 @@
+import ImpactTdContents from './ImpactTdContents';
+import ImpactThContents from './ImpactThContents';
+
+export { ImpactThContents, ImpactTdContents };
